LP-575:
Yo como moderador quiero rechazar o aceptar registros reportados para evitar uso indebido de la aplicación.

+++ source/LoCoMPro/ViewModels/Cuenta/AporteVM.cs
@@ -32,6 +32,6 @@ namespace LoCoMPro.ViewModels.Cuenta
         public required string canton { get; set; }
         // Calificación
         [Display(Name = "Calificación")]
-        public decimal calificacion { get; set; }
+        public double calificacion { get; set; }
     }
 }
